carver 层：小屋 Y 改用挖后高度图，全版本启用
根因：SwampHutPiece.postProcess 在 FEATURES 阶段读 LevelAccessor.getHeightmapPos
(MOTION_BLOCKING_NO_LEAVES)，而 carver 在更早的 CARVERS 阶段；FEATURES 开头的
primeHeightmaps 会从实时方块重扫高度图。我们原先只算密度加含水层（即挖前），
因此凡 footprint 有地表洞穴口的候选 Y 一律偏高。

- 新增 carver.c/carver.h：复刻 NoiseBasedChunkGenerator.applyCarvers 的
  17x17 origin x carverIndex 遍历，共享 target 的 CarvingMask，写入夹取到 target
  重叠区；含 Cave/CanyonWorldCarver、carveEllipsoid/carveBlock、Mth.sin/cos。
- 新增 sin_table.h：由 JVM dump 的 Mth.SIN（65536 项）。该表依 JVM/平台而异，
  与 vanilla 自身的表同理，已在文件头注明。
- phase2：per-chunk 雕刻缓存 + carved-aware 高度；保留 pre-carver 高度供群系门用。
- eval：carver 对所有版本开启。播种在 1.18.2~26.1.2 之间逐指令相同（无 |1L），
  configured_carver 配置对 1.20+ 逐字段相同；1.18.2/1.19.2 的配置本地无法验证，
  已在源码注明。
- rng：修复 lysh_lcg_next_long 未对第二个 next(32) 做符号扩展（Java 为 i2l）。
  它影响 setLargeFeatureSeed，即小屋朝向 dir 与 carver 播种；因 footprint 形状
  只取决于 dir 的奇偶，该缺陷此前一直静默。
- 灌水优先判定 + 守卫（阈值 64）：灌水判定与雕刻无关，挖前已灌满的直接拒并跳过
  雕刻层。Y 门不可用于跳过 —— 雕刻只会降低列顶，挖前超门槛的候选挖后可能合格。
- 阶段 2 性能：插值 cell 缓存由单 cell 改为 per-chunk 5x5 角点列网格，
  lysh scan 100M 格阶段 2 由 137.65 ms/候选降至约 7.5 ms/候选。
  多种子（2048 种子、默认范围）边际 8.87 ms/种子（开）对 8.88（关），份额约 0。

文档：lysh-c/README.md 补回丢失的 §4.2（11 个小屋回归网）与新增 §2.5.5（雕刻层）；
修正 setCarverSeed 的旧说法（26.1.2 不存在，任何版本无 |1L）；性能数字更新；
删除叙事性案例。根 README 的 C 内核条目补一句雕刻层。

阶段 1 未改动。

+++ lysh-c/src/rng.c
@@ -195,9 +195,18 @@ int lysh_lcg_next_bits(lysh_lcg *r, int bits) {
 }
 
 uint64_t lysh_lcg_next_long(lysh_lcg *r) {
+    /* ⚠️ java.util.Random.nextLong()（以及 BitRandomSource.nextLong）是
+     *     return ((long)next(32) << 32) + next(32);
+     * 第二个 next(32) 是 **i2l**（符号扩展）再加，不是零扩展！
+     * 所以当第二个 next(32) 的最高位为 1 时，结果会比"拼接"小 2^32。
+     * 26.1.2 的字节码逐条（BitRandomSource.nextLong）：
+     *     next(32) -> istore_1 ; next(32) -> istore_2 ;
+     *     iload_1; i2l; bipush 32; lshl -> lstore_3 ; lload_3; iload_2; i2l; ladd
+     * 这里错一位就会让 `setCarverSeed` 派生出的 LCG 状态整体跑偏，
+     * 从而让**小屋朝向 dir** 与**雕刻器播种**都错（实测：本种子下 ci=2 的 b 差 2^32）。 */
     uint64_t hi = (uint64_t)(uint32_t)lysh_lcg_next_bits(r, 32);
-    uint64_t lo = (uint64_t)(uint32_t)lysh_lcg_next_bits(r, 32);
-    return (hi << 32) + lo;
+    int64_t lo = (int64_t)lysh_lcg_next_bits(r, 32);      /* 符号扩展 */
+    return (hi << 32) + (uint64_t)lo;
 }
 
 /* JDK 的 Random.nextInt(bound) */
